Renamed Core->Common (namespaces, project, usings)
