+++ source/Conference/Common/ICommand.cs
@@ -10,10 +10,9 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-namespace Core
+namespace Common
 {
-    public interface ICommandBus
+    public interface ICommand
     {
-        void Send(ICommand command);
     }
 }
