+++ source/Conference/Common/ICommandBus.cs
@@ -10,9 +10,10 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-namespace Core
+namespace Common
 {
-    public interface IEvent
+    public interface ICommandBus
     {
+        void Send(ICommand command);
     }
 }
