+++ source/Conference/Common/IEvent.cs
@@ -10,9 +10,9 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-namespace Core
+namespace Common
 {
-    public interface ICommand
+    public interface IEvent
     {
     }
 }
