+++ source/Conference/Common/IRepository.cs
@@ -10,10 +10,10 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-using System;
-
-namespace Registration
+namespace Common
 {
+    using System;
+
     public interface IRepository
     {
         T Find<T>(Guid id) where T : class, IAggregateRoot;
