+++ source/Conference/Common/Properties/AssemblyInfo.cs
