+++ source/Conference/Registration/ConferenceSeatsAvailability.cs
@@ -10,13 +10,14 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-
 namespace Registration
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Collections.ObjectModel;
+    using System.Linq;
+    using Common;
+
     public class ConferenceSeatsAvailability : IAggregateRoot
     {
         // ORM requirement
