+++ source/Conference/Registration/Reservation.cs
@@ -10,11 +10,11 @@
 // See the License for the specific language governing permissions and limitations under the License.
 // ==============================================================================================================
 
-using System;
-using System.ComponentModel.DataAnnotations;
-
 namespace Registration
 {
+    using System;
+    using System.ComponentModel.DataAnnotations;
+
     [ComplexType]
     public class Reservation
     {
